Move communication stuff to communication.js
