+++ src/app/bug-reporting/metrics/communication.js
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2017 The "mysteriumnetwork/mysterium-vpn" Authors.
+ * Copyright (C) 2018 The "MysteriumNetwork/mysterium-vpn" Authors.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -17,12 +17,14 @@
 
 // @flow
 
-export type MapSyncDTO<T: string> = {
+type MapSyncDTO<T: string> = {
   metric: T,
   value: mixed
 }
 
-export interface MapSyncCommunication<T: string> {
+interface Communication<T: string> {
   sendMapUpdate (update: MapSyncDTO<T>): void,
   onMapUpdate (callback: (MapSyncDTO<T>) => void): void
 }
+
+export type { MapSyncDTO, Communication }
